requirements.txt
openai-whisper
torch
deep-translator
srt
ffmpeg-python
argostranslate